Update circle meter descriptions and samples
