--- src/components/CircleMeter/CircleMeter.md
@@ -1,22 +1,0 @@
-```js
-<div>
-  <CircleMeter value={35} size={150} style={{ margin: '0 20px' }} />
-  <CircleMeter
-    value={35}
-    size={150}
-    progressStart={230}
-    meterLength={260}
-    rounded={true}
-    style={{ margin: '0 20px' }}
-  />
-  <CircleMeter value={35} size={150} progressStart={180} thickness={36} style={{ margin: '0 20px' }} />
-  <CircleMeter
-    value={35}
-    size={150}
-    progressStart={180}
-    meterLength={240}
-    textStyle={{ fill: 'black', fontSize: '40px' }}
-    style={{ margin: '0 20px' }}
-  />
-</div>
-```
