+++ src/docs/examples/BlockMeter/BlockMeterExample.js
